[Automatic Import] Refactor useEffect to reduce complexity (#194150)
(cherry picked from commit bcdb0d8ede90a99317bddd99cc122dc98107d1a7)
